Support luarocks/loverocks
This adds a rockspec file which can be used to build and install gifcat
like so:

    loverocks install <url to gifcat-scm-1.rockspec>

Or, once it's uploaded to https://luarocks.org:

    loverocks install gifcat

To support this, I had to move the threaded code into a string. Guessing
the filename from the module name isn't reliable enough: It will only
work when the path it's being required from is exactly `./?.lua`.

Also fixed a typo in the example.